Add update to change_branch to decrease dependency

+++ app/controllers/repositories_controller.rb
@@ -26,12 +26,14 @@ class RepositoriesController < ApplicationController
   end
 
   def change_branch
-    repository = Repository.find(params[:repository].to_i)
-    delete_repo(repository.name)
-    repository.destroy
-    branches = list_branches(@repository)
-    saved_repository = Repository.save_repository_to_db(params[:username], params[:repository], params[:uid], session[:user_id], branches)
-    rows = CodeReview.new(repository, params[:username], params[:branch])
+    if repository = Repository.find(params[:repository].to_i)
+      repo_name = repository.name
+      delete_repo(repository.name)
+      repository.destroy
+    end
+    rows = CodeReview.new(repo_name, params[:username], params[:branch]).rows
+    branches = list_branches(repo_name)
+    saved_repository = Repository.save_repository_to_db(params[:username], repo_name, params[:uid], session[:user_id], branches)
     rows.map do |repo_file|
       RepositoryFile.create_repo_files(repo_file, params[:username], saved_repository)
     end
